+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [v1.1](https://github.com/d8corp/async-mobx/tree/v1.1)
+### v1.1.0 [![18.10.2020](https://img.shields.io/date/1603029826)](https://github.com/d8corp/async-mobx/tree/v1.1.0)
+- added `reset` method
+- improved typing
+- added options `keepResponse` and `keepError`
+- removed value from listeners
+- then, catch and finally return a promise
+
 ## [v1.0](https://github.com/d8corp/async-mobx/tree/v1.0)
 ### v1.0.1 [![03.10.2020](https://img.shields.io/date/1601738568)](https://github.com/d8corp/async-mobx/tree/v1.0.1)
 - added type of value